feat: 게임 로직 함수 구현 및 열거형 타입 분리,구현, 가위바위보 게임 함수 수정.
