+++ RockPaperScissors/RockPaperScissors/GameResult.swift
@@ -1,0 +1,12 @@
+//
+//  GameResult.swift
+//  RockPaperScissors
+//
+//  Created by Minseong Kang on 2023/05/05.
+//
+
+enum GameResult {
+	case win
+	case draw
+	case lose
+}
