+++ RockPaperScissors/RockPaperScissors/GameTurn.swift
@@ -1,0 +1,13 @@
+//
+//  GameTurn.swift
+//  RockPaperScissors
+//
+//  Created by Minseong Kang on 2023/05/05.
+//
+
+enum GameTurns {
+	case user
+	case computer
+	case gameEnd
+}
+
