+++ RockPaperScissors/RockPaperScissors/HandShape.swift
@@ -1,0 +1,12 @@
+//
+//  HandShape.swift
+//  RockPaperScissors
+//
+//  Created by Minseong Kang on 2023/05/05.
+//
+
+enum HandShape: CaseIterable {
+	case scissors
+	case rock
+	case paper
+}
